fix: refactor template rendering logic, abstract common func
