+++ pkg/utils/strings/strings.go
@@ -17,10 +17,17 @@ limitations under the License.
 package strings
 
 import (
+	"errors"
 	"fmt"
+	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 	"unicode"
+
+	"github.com/labring/sealos/pkg/constants"
+	"github.com/labring/sealos/pkg/template"
+	"github.com/labring/sealos/pkg/utils/file"
 
 	"golang.org/x/exp/slices"
 
@@ -156,6 +163,67 @@ func RenderTextWithEnv(text string, envs map[string]string) string {
 		text = strings.ReplaceAll(text, o, n)
 	}
 	return text
+}
+
+func RenderTemplatesWithEnv(filePaths string, envs map[string]string) error {
+	var (
+		renderEtc       = filepath.Join(filePaths, constants.EtcDirName)
+		renderScripts   = filepath.Join(filePaths, constants.ScriptsDirName)
+		renderManifests = filepath.Join(filePaths, constants.ManifestsDirName)
+	)
+
+	for _, dir := range []string{renderEtc, renderScripts, renderManifests} {
+		logger.Debug("render env dir: %s", dir)
+		if !file.IsExist(dir) {
+			logger.Debug("Directory %s does not exist, skipping", dir)
+			continue
+		}
+
+		if err := filepath.Walk(dir, func(path string, info os.FileInfo, errIn error) error {
+			if errIn != nil {
+				return errIn
+			}
+			if info.IsDir() || !strings.HasSuffix(info.Name(), constants.TemplateSuffix) {
+				return nil
+			}
+
+			fileName := strings.TrimSuffix(path, constants.TemplateSuffix)
+			if file.IsExist(fileName) {
+				if err := os.Remove(fileName); err != nil {
+					logger.Warn("failed to remove existing file [%s]: %v", fileName, err)
+				}
+			}
+
+			writer, err := os.OpenFile(fileName, os.O_CREATE|os.O_RDWR, os.ModePerm)
+			if err != nil {
+				return fmt.Errorf("failed to open file [%s] for rendering: %v", path, err)
+			}
+			defer writer.Close()
+
+			body, err := file.ReadAll(path)
+			if err != nil {
+				return err
+			}
+
+			t, isOk, err := template.TryParse(string(body))
+			if isOk {
+				if err != nil {
+					return fmt.Errorf("failed to create template: %s %v", path, err)
+				}
+				if err := t.Execute(writer, envs); err != nil {
+					return fmt.Errorf("failed to render env template: %s %v", path, err)
+				}
+			} else {
+				return errors.New("parse template failed")
+			}
+
+			return nil
+		}); err != nil {
+			return fmt.Errorf("failed to render templates in directory %s: %v", dir, err)
+		}
+	}
+
+	return nil
 }
 
 func TrimQuotes(s string) string {
